Se termino de implementar la seguridad con JWT

+++ src/controllers/ProjectController.ts
@@ -17,7 +17,11 @@ export class ProjectController {
 
     static getAllProjects = async (req : Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [ //Con esta condición se le dice que traiga nada más los proyectos en el cual el manager creó. 
+                    {manager: {$in: req.user._id}}
+                ]
+            })
             res.json(projects)
         } catch (error) {
             console.log(error)
@@ -31,6 +35,11 @@ export class ProjectController {
             const project = await Project.findById(id).populate('tasks')
             if (!project) {
                 res.status(404).json({ error: 'Project not found' })
+                return
+            }
+
+            if(project.manager.toString() !== req.user._id.toString()){
+                res.status(404).json({ error: 'Action no validate'})
                 return
             }
             res.json(project)
@@ -48,6 +57,12 @@ export class ProjectController {
                 res.status(404).json({ error: 'Project not found' })
                 return
             }
+
+            if(project.manager.toString() !== req.user._id.toString()){
+                res.status(404).json({ error: 'Only the manager can update this project.'})
+                return
+            }
+
             project.client_name = req.body.client_name
             project.project_name = req.body.project_name
             project.description = req.body.description
@@ -67,6 +82,12 @@ export class ProjectController {
                 res.status(404).json({ error: 'Project not found' })
                 return
             }
+
+            if(project.manager.toString() !== req.user._id.toString()){
+                res.status(404).json({ error: 'Only the manager can delete this project.'})
+                return
+            }
+            
             await project.deleteOne()
             res.send('Project deleted succesfully')
         } catch (error) {
